requirements.txt
azure-cognitiveservices-speech
gradio
openai
streamlit
git+https://github.com/paulocoutinhox/st-message
